refactor: remove unnecessary folder nesting
