+++ src/pages/MainPage.tsx
@@ -1,10 +1,10 @@
 import { useState } from 'react';
-import Modal from '../components/Modal/Modal';
-import Button from '../components/Button/Button';
-import UncontrolledForm from '../components/Forms/UncontrolledForm/UncontrolledForm';
-import ControlledForm from '../components/Forms/ControlledForm/ControlledForm';
+import Modal from '../components/Modal';
+import Button from '../components/Button';
+import UncontrolledForm from '../components/Forms/UncontrolledForm';
+import ControlledForm from '../components/Forms/ControlledForm';
 import HeroImage from '../components/HeroImage';
-import Layout from '../components/Layout/Layout';
+import Layout from '../components/Layout';
 import handleFormSubmission from '../utils/handleFormSubmission';
 import { useAppSelector } from '../store/hooks';
 import { selectSubmissions } from '../store/formSelectors';
