+++ src/tests/Button.test.tsx
@@ -1,4 +1,4 @@
-import Button from '../components/Button/Button';
+import Button from '../components/Button';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
