+++ src/tests/Modal.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { describe, expect, vi } from 'vitest';
-import Modal from '../components/Modal/Modal';
+import Modal from '../components/Modal';
 import userEvent from '@testing-library/user-event';
 
 describe('Modal', () => {
