+++ src/tests/UncontrolledForm.test.tsx
@@ -1,7 +1,7 @@
 import { describe, expect, vi, beforeEach } from 'vitest';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import UncontrolledForm from '../components/Forms/UncontrolledForm/UncontrolledForm';
+import UncontrolledForm from '../components/Forms/UncontrolledForm';
 import * as handleSubmitModule from '../utils/handleUncontrolledFormSubmit';
 
 describe('UncontrolledForm', () => {
